remove spectroscopyGraph call from the ITC
