+++ itc/client/shared/src/main/scala/lucuma/itc/client/queries.scala
@@ -187,95 +187,6 @@ object ImagingIntegrationTime extends GraphQLOperation[Unit] {
     (c: HCursor) => c.downField("imaging").as[ClientCalculationResult]
 }
 
-object SpectroscopyGraphsQuery extends GraphQLOperation[Unit] {
-  type Data      = SpectroscopyGraphsResult
-  type Variables = SpectroscopyGraphsInput
-
-  val document =
-    """
-      query($input: SpectroscopyGraphsInput!) {
-        spectroscopyGraphs(input: $input) {
-          versions {
-            serverVersion
-            dataVersion
-          }
-          targetGraphs {
-            ... on TargetGraphsResult {
-              graphs {
-                peakFinalSNRatio
-                atWavelengthFinalSNRatio
-                peakSingleSNRatio
-                atWavelengthSingleSNRatio
-                ccds {
-                  singleSNRatio
-                  totalSNRatio
-                  peakPixelFlux
-                  ampGain
-                  maxTotalSNRatio
-                  maxSingleSNRatio
-                  wavelengthForMaxTotalSNRatio {
-                    picometers
-                  }
-                  wavelengthForMaxSingleSNRatio {
-                    picometers
-                  }
-                  wellDepth
-                  warnings {
-                    msg
-                  }
-                }
-                graphData {
-                  graphType
-                  series {
-                    title
-                    seriesType
-                    dataY
-                    xAxis {
-                      start
-                      end
-                      count
-                      min
-                      max
-                    }
-                    yAxis {
-                      start
-                      end
-                      count
-                      min
-                      max
-                    }
-                  }
-                }
-              }
-              band
-              emissionLine { picometers }
-            }
-            ... on TargetError {
-              errorCode
-              message
-              wellHalfFilledSeconds
-              wavelength { picometers }
-            }
-          }
-        }
-      }
-  """
-
-  override val varEncoder: Encoder.AsObject[Variables] =
-    Encoder.AsObject.instance[SpectroscopyGraphsInput] { input =>
-      JsonObject(
-        "input" -> Encoder[SpectroscopyGraphsInput].apply(input)
-      )
-    }
-
-  override val dataDecoder: Decoder[SpectroscopyGraphsResult] = c =>
-    val result = c.downField("spectroscopyGraphs")
-    for
-      versions     <- result.downField("versions").as[ItcVersions]
-      targetGraphs <- result.downField("targetGraphs").as[AsterismTargetGraphsResultOutcomes]
-    yield SpectroscopyGraphsResult(versions, targetGraphs)
-}
-
 object SpectroscopyIntegrationTimeAndGraphsQuery extends GraphQLOperation[Unit] {
   type Data      = SpectroscopyIntegrationTimeAndGraphsResult
   type Variables = SpectroscopyIntegrationTimeAndGraphsInput
