+++ itc/service/src/main/resources/graphql/itc_base.graphql
@@ -3,7 +3,6 @@
 ITC schema
 ##########
 """
-
 type Query {
   versions: ItcVersions
 
@@ -22,14 +21,6 @@ type Query {
   imaging(
     input: ImagingInput!
   ): CalculationResult!
-
-  """
-  Given an instrument and a mode, return the graph data for the given targets.
-  Values are rounded to limit the size of the response.
-  """
-  spectroscopyGraphs(
-    input: SpectroscopyGraphsInput!
-  ): SpectroscopyGraphsResult!
 
   """
   Given an instrument and a mode, return the integration time,
@@ -311,30 +302,6 @@ input SignificantFiguresInput {
   ccd: PosInt
 }
 
-"""Params for spectroscopy graphs calculation"""
-input SpectroscopyGraphsInput {
-  """Observing wavelength."""
-  atWavelength: WavelengthInput!
-
-  """Exposure time duration"""
-  exposureTime: TimeSpanInput!
-
-  """Exposures"""
-  exposureCount: PosInt!
-
-  """Relevant data for targets in asterism."""
-  asterism: [TargetInput!]!
-
-  """Conditions"""
-  constraints: ItcConstraintsInput!
-
-  """Instrument modes"""
-  mode: InstrumentModesInput!
-
-  """Significant figures, truncates the amount of significant value"""
-  significantFigures: SignificantFiguresInput
-}
-
 """Parameters to retrieve graph data"""
 input SpectroscopyIntegrationTimeAndGraphsInput {
 
@@ -405,7 +372,7 @@ type TargetError {
   wavelength: Wavelength
 }
 
-"""Graph data types"""
+"""Graph series types"""
 enum ItcSeriesType {
   SIGNAL_DATA
 
@@ -467,7 +434,9 @@ type ItcSeries {
 }
 
 type ItcGraph {
-  """Data type: alternatives SignalGraph, S2NGraph"""
+  """
+  Data type: alternatives SignalGraph, S2NGraph
+  """
   graphType: ItcGraphType!
 
   series: [ItcSeries!]!
@@ -528,27 +497,6 @@ type TargetGraphs {
   """SN Ratio at the requested wavelength if passed along"""
   atWavelengthSingleSNRatio: SignalToNoise
 }
-
-type TargetGraphsResult {
-  """graphs for the target"""
-  graphs: TargetGraphs!
-
-  """The band used for the graphs calculation, in case of a BandNormalized SED"""
-  band: Band
-
-  """The emission line used for the time calculation, in case of an EmissionLines SED"""
-  emissionLine: Wavelength
-}
-
-type SpectroscopyGraphsResult {
-  """server and data versions"""
-  versions: ItcVersions!
-
-  """results for each target, in same order as input"""
-  targetGraphs: [TargetGraphsOutcome!]!
-}
-
-union TargetGraphsOutcome = TargetGraphsResult | TargetError
 
 type SpectroscopyTimeAndGraphsResult {
   """server and data versions"""
