+++ itc/service/src/main/scala/lucuma/itc/Itc.scala
@@ -3,11 +3,7 @@
 
 package lucuma.itc.service
 
-import eu.timepit.refined.types.numeric.PosInt
-import lucuma.core.math.Wavelength
 import lucuma.core.model.ExposureTimeMode
-import lucuma.core.util.TimeSpan
-import lucuma.itc.TargetGraphsCalcResult
 import lucuma.itc.TargetIntegrationTime
 import lucuma.itc.TargetTimeAndGraphs
 
@@ -22,18 +18,6 @@ trait Itc[F[_]]:
     constraints:      ItcObservingConditions,
     exposureTimeMode: ExposureTimeMode
   ): F[TargetIntegrationTime]
-
-  /**
-   * Retrieve the graph data for the given mode and exposureTime and exposures
-   */
-  def calculateGraphs(
-    target:        TargetData,
-    atWavelength:  Wavelength,
-    observingMode: ObservingMode,
-    constraints:   ItcObservingConditions,
-    exposureTime:  TimeSpan,
-    exposureCount: PosInt
-  ): F[TargetGraphsCalcResult]
 
   def calculateTimeAndGraphs(
     target:           TargetData,
